requirements.txt
alpaca-trade-api
pandas
pandas-ta
python-dotenv